Update customize.md
The values.yaml file uses the key `replica` instead of `replicas`. This causes the template value of 3 to not be inserted into the deployments. I don't know if this was intentional, since the next page in the tutorial asks the user to verify the dry-run output is correct. But since that page doesn't call out the omission, I figured it was unintentional and should probably be corrected here (or perhaps better to be called out on the next page).