Minor doc enhancements to GitOps module (#440)
* Tweaked command to check for the presence of helm

When `helm` is already installed, the `helm ls` command might not return anything. This can be confusing to the reader.  

Changed to execute the `helm version` command so that something is always displayed: either the version or a "command not found" error is displayed.

* Clarified that "helm ls" may return nothing

Under some normal circumstances, "helm ls" may be expected to return nothing.  Doc has been updated to set that expectation. Otherwise, users may be expecting this command to return some data.

* Change to more conventional clone command

Aligns with other examples in the workshop and is the approach recommended by GitHub.com. use of "git clone git@github.com:" failed in the context of how my GitHub.com account is set up:

```
$ git clone git@github.com:-----/eks-example.git
Cloning into 'eks-example'...
Warning: Permanently added the RSA host key for IP address '140.82.114.3' to the list of known hosts.
Permission denied (publickey).
fatal: Could not read from remote repository.
```

*  Change to more conventional clone command

Align with change made to `installweaveflux.md`.