Update connect.md
This command works for me when the original command did not.  Also, the command will now match what is in the screenshot - which is where I got the command from.

This might need further testing as my cluster had a different name.  Also, when running 
    $ aws eks get-token --cluster-name eksworkshop-eksctl | jq -r '.status.token'
different values are returned if different regions are specified with for example '--region us-west-1' after the 'aws' command.