Added Module for KMS and K8s Secrets. eksctl 0.15.0 Dependency
This module is dependent on eksctl 0.15.0. It should not be merged until it is released.
Updated the workshop prerequisites to include creation of a KMS CMK
Updated the cluster deployment to use a config file rather than the CLI
Added a module to step the attendee through using secrets with KMS
Added relevant screenshots.