Option to increase Cloud9 disk size
I had issues installing Tensorflow after running several of the modules on Cloud9. The reason behind that failure was the lack of space on the Cloud9 environment.

I copied the script from ECS Workshop, I think is very elegant and simple way to increase the size of the disk until Cloud9 allow to add an additional disk from creation.