Removed unnecessary fields from fluentbit.yaml (#1008)
Since the master node is not visible to the user in EKS, there is no need for tolerations settings to prevent fluentbit from being deployed to the master node.
The environment variables FLUENT_ELASTICSEARCH_HOST and FLUENT_ELASTICSEARCH_PORT are not used. These are unnecessary because they are defined directly in output-elasticsearch.conf in ConfigMap.