Update basic-pod-limits.md
Under “BASIC POD CPU AND MEMORY MANAGEMENT“ phase, the first pod that our customers need to deploy scales to 2G memory.
“kubectl run --requests=memory=1G,cpu=0.5 --image hande007/stress-ng basic-request-pod --restart=Never -- --vm-keep --vm-bytes 2g --timeout 600s --vm 1 --oomable –verbose”
In the prerequisites of the workshop we provide the customers yaml that build the EKS with t3.small instances which has 2g ram per node. Since that we cannot run 2g memory containers under these nodes as shown in the workshop documentation. I changed the pod to use only 1.5 in order to get the right output